Перенесла кусок графики уровня в отдельный класс
